fix: 장학금 신청 기능 구현 중 발생한 MyBatis 연결 에러 해결

+++ src/main/java/com/csee/swplus/mileage/scholarship/mapper/ScholarshipMapper.java
@@ -7,7 +7,7 @@ import java.time.LocalDateTime;
 
 @Mapper
 public interface ScholarshipMapper {
-    int createApplication(@Param("studentId") Long studentId,
+    int createApplication(@Param("studentId") int studentId,
                           @Param("applyDate") LocalDateTime applyDate,
                           @Param("isChecked") int isChecked);
 }
